% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfda.R
\name{fdagg}
\alias{fdagg}
\title{This function plots the FECs}
\usage{
fdagg(fdobj, ylab = "Value", xlab = "Time", addzero = FALSE, title = "",
  gridout = TRUE)
}
\arguments{
\item{fdobj}{FECs}
}
\description{
This function plots the FECs
}
\examples{
\dontrun{
}
}
