% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEC_conditional.R
\name{FECconditional}
\alias{FECconditional}
\title{This function computes the next FEC conditional on previous fitted FEC using the SOAP method}
\usage{
FECconditional(ylist, tlist, beta_previous, spline_basis, threshold = 1e-04,
  gamma = 0)
}
\arguments{
\item{ylist}{a list of observed longtitudinal data; each element in the list is a numeric vector for one subject's observed data}

\item{tlist}{a list of observed time points}

\item{beta_previous}{a list of previously estimated FECs' coefficients}

\item{spline_basis}{the B-spline basis for the fitted FEC}

\item{gamma}{a positive number, the smoothing parameter}
}
\description{
This function computes the next FEC conditional on previous fitted FEC using the SOAP method
}
\examples{
\dontrun{
library(fda)
library(cFuSIM)
data(bike_cFuSIM)
timepts = bike$timepts
norder=4 ## cubic B-spline
nbasis=norder+length(timepts)-2; 
spline_basis=create.bspline.basis(rangeval=c(1,24),nbasis#'norder,timepts)
wull = bike$temp
xfds=  Data2fd(y=wull\%>\%t, argvals=bike$timepts)
y = bike$y
train_sample = 1:length(y)
y = y[train_sample]
xfd = xfds[train_sample]
res_c = cFuSIM_index(y, xfd, spline_basis)
beta_fd = fd(res_c$coefBeta, res_c$basisBeta)
plot(beta_fd,ylab="index function", xlab='time')
fdagg(beta_fd)
score_fit = (res_c$score_fit)
pred_y = localpoly.reg(score_fit, y, degree.pol = 1, kernel.type = "gaussian",bandwidth = "CV",deriv=0,points#'score_fit)
plot(x=score_fit, y=y)
lines(pred_y$predicted[order(score_fit)],x=score_fit[order#'score_fit)],col=4)
}
}
